test: moved dto tests to correct folder
