+++ packages/figma_variables_api/test/src/dto/variable_mode_value/variable_mode_value_dto_test.dart
